requirements.txt
opencv-python
librosa
python-dotenv
pandas
matplotlib
seaborn
scikit-learn